simplify test file layout
